move domain loader tests
