+++ api-js/src/loaders/domains/__tests__/load-domain-by-domain.test.js
@@ -3,10 +3,10 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { domainLoaderByDomain } = require('../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { domainLoaderByDomain } = require('../..')
 
 describe('given a domainLoaderByDomain dataloader', () => {
   let query, drop, truncate, migrate, collections, i18n
