+++ api-js/src/loaders/domains/__tests__/load-domain-connections-by-user-id.test.js
@@ -5,14 +5,14 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
 const {
   domainLoaderConnectionsByUserId,
   domainLoaderByKey,
-} = require('../loaders')
+} = require('../..')
 
 describe('given the load domain connections by user id function', () => {
   let query,
